+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 1.2.38 (unreleased)
 -------------------
+* load default meta_types for new meeting [mihaitab]
 
 1.2.37 (2013-03-26)
 -------------------